Move instructions section to its own class
